domain-intelligence-engine/requirements.txt
anthropic>=0.40.0
youtube-transcript-api>=0.6.0
yt-dlp>=2024.0.0
supabase>=2.0.0
openai>=1.0.0
streamlit>=1.30.0
python-dotenv>=1.0.0
pyyaml>=6.0
tiktoken>=0.5.0
numpy>=1.24.0
